requirements.txt
requests>=2,<3
python-decouple==3.4